refactor: 학교 정보 불러오기 타입 수정
